Move files that will be generated to the modflow folder
