+++ flopy4/mf6/modflow/gwf/__init__.py
@@ -1,0 +1,8 @@
+from flopy4.mf6.modflow.gwf.chd import Chd
+from flopy4.mf6.modflow.gwf.dis import Dis
+from flopy4.mf6.modflow.gwf.gwf import Gwf
+from flopy4.mf6.modflow.gwf.ic import Ic
+from flopy4.mf6.modflow.gwf.npf import Npf
+from flopy4.mf6.modflow.gwf.oc import Oc
+
+__all__ = ["Gwf", "Chd", "Dis", "Ic", "Npf", "Oc"]
