+++ test/test_component.py
@@ -9,10 +9,10 @@ from xarray import DataTree
 
 from flopy4.mf6.component import COMPONENTS
 from flopy4.mf6.constants import FILL_DNODATA
-from flopy4.mf6.gwf import Chd, Dis, Gwf, Ic, Npf, Oc
-from flopy4.mf6.ims import Ims
-from flopy4.mf6.simulation import Simulation
-from flopy4.mf6.tdis import Tdis
+from flopy4.mf6.modflow.gwf import Chd, Dis, Gwf, Ic, Npf, Oc
+from flopy4.mf6.modflow.ims import Ims
+from flopy4.mf6.modflow.simulation import Simulation
+from flopy4.mf6.modflow.tdis import Tdis
 
 
 def test_registry():
