+++ test/test_spec.py
@@ -1,4 +1,4 @@
-from flopy4.mf6.gwf import Gwf
+from flopy4.mf6.modflow.gwf import Gwf
 from flopy4.mf6.spec import blocks, blocks_dict
